Initial integration between RAUC and Aktualizr
Install some extra helper scripts along with RAUC; those scripts are
responsible for the integration with Aktualizr. Also, modify the
system.conf file adding a "handlers" section that refers to one of the
helper scripts.

Notice this is a PoC integration just for tests; calc-root-hash.sh
probably doesn't belong here.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>